エラーハンドリング？追加

+++ app/controllers/api/tasks_controller.rb
@@ -2,33 +2,54 @@ class Api::TasksController < ApplicationController
   protect_from_forgery with: :null_session
   def index
     tasks = Task.all
-    render json: tasks
+    if tasks.empty?
+      render json: { status: "OK", result: "タスクはありません" }, status: :ok
+      return
+    end
+    render json: { status: "OK", result: tasks }, status: :ok
   end
 
   def show
-    task = Task.find(params[:id])
-    render json: task
+    task = Task.find_by(id: params[:id])
+    id =  params[:id]
+    if task.nil?
+      render json: { status: "OK", result: "[ID:#{id}]そのタスクは見つかりません" }, status: :ok
+      return
+    end
+    render json: { status: "OK", result: task }, status: :ok
   end
 
   def create
-    Task.create(task_params)
-    render json: { status: "OK" }, status: :ok
+    task = Task.create(task_params)
+    render json: { status: "OK", result: task }, status: :ok
   end
 
   def update
-    task = Task.find(params[:id])
+    task = Task.find_by(id: params[:id])
+    if task.nil?
+      render json: { status: "OK", result: "[ID:#{params[:id]}]更新するタスクが見つかりません" }
+      return
+    end
     task.update(task_params)
-    render json: { status: "OK" }, status: :ok
+    render json: { status: "OK", result: task }, status: :ok
   end
 
   def destroy
-    task = Task.find(params[:id])
-    task.destroy
-    render json: { status: "OK" }, status: :ok
+    task = Task.find_by(id: params[:id])
+    if task.nil?
+      render json: { status: "OK", result: "[ID:#{params[:id]}] 削除するタスクが見つかりません" }
+      return
+    end
+    task = task.destroy
+    render json: { status: "OK", result: task }, status: :ok
   end
 
   def delete_recently_task
     task = Task.order(created_at: :desc).first
+    if task.nil?
+      render json: { status: "OK", result: "タスクが1つもありません" }, status: :ok
+      return
+    end
     result = task.destroy
     render json: result, status: :ok
   end
